face_attendance/requirements.txt
flask
dlib
face_recognition
pillow
imutils
numpy
opencv-python
h5py
sklearn